hackathon-idea-generator/requirements.txt
google-generativeai==0.3.2
streamlit==1.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
